mev/requirements.txt
asgiref==3.5.2
backoff==1.10.0
billiard==3.6.4.0
boto3==1.24.54
cached-property==1.5.2
cachetools==4.2.2
celery==5.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==4.0.7
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-storages==1.13.1
docker==5.0.0
et-xmlfile==1.1.0
future==0.18.2
ghp-import==2.0.1
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.1
google-cloud-storage==1.39.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.6.0
install==1.3.4
Jinja2==3.0.1
joblib==1.2.0
livereload==2.6.3
lunr==0.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkautodoc==0.1.0
mkdocs==1.2.3
nltk==3.6.6
numpy==1.23.0
oauthlib==3.2.1
openpyxl==3.0.7
packaging==20.9
pandas==1.4.3
prompt-toolkit==3.0.19
protobuf==3.18.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyyaml-env-tag==0.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.9.2
sentry-sdk==1.5.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tables==3.6.1
tornado==6.1
tqdm==4.61.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
watchdog==2.1.3
wcwidth==0.2.5
websocket-client==1.1.0
xlrd==2.0.1
zipp==3.4.1
